+++ app/src/main/java/org/simple/clinic/onboarding/OnboardingModule.kt
@@ -4,12 +4,13 @@ import com.f2prateek.rx.preferences2.Preference
 import com.f2prateek.rx.preferences2.RxSharedPreferences
 import dagger.Module
 import dagger.Provides
-import javax.inject.Named
+import org.simple.clinic.main.TypedPreference
+import org.simple.clinic.main.TypedPreference.Type.OnboardingComplete
 
 @Module
 class OnboardingModule {
 
   @Provides
-  @Named("onboarding_complete")
+  @TypedPreference(OnboardingComplete)
   fun hasTheUserCompletedOnboarding(rxSharedPreferences: RxSharedPreferences): Preference<Boolean> = rxSharedPreferences.getBoolean("onboarding_complete")
 }