+++ app/src/main/java/org/simple/clinic/setup/SetupActivityEffect.kt
@@ -13,3 +13,7 @@ object InitializeDatabase : SetupActivityEffect()
 object ShowCountrySelectionScreen : SetupActivityEffect()
 
 object SetFallbackCountryAsCurrentCountry: SetupActivityEffect()
+
+object RunDatabaseMaintenance: SetupActivityEffect()
+
+object FetchDatabaseMaintenanceLastRunAtTime: SetupActivityEffect()