+++ app/src/main/java/org/simple/clinic/setup/SetupActivityModule.kt
@@ -1,24 +1,47 @@
 package org.simple.clinic.setup
 
+import com.f2prateek.rx.preferences2.Preference
+import com.f2prateek.rx.preferences2.RxSharedPreferences
 import dagger.Module
 import dagger.Provides
 import org.simple.clinic.BuildConfig
 import org.simple.clinic.appconfig.Country
 import org.simple.clinic.di.AssistedInjectModule
+import org.simple.clinic.main.TypedPreference
+import org.simple.clinic.main.TypedPreference.Type.DatabaseMaintenanceRunAt
+import org.simple.clinic.main.TypedPreference.Type.FallbackCountry
+import org.simple.clinic.util.Optional
+import org.simple.clinic.util.preference.InstantRxPreferencesConverter
+import org.simple.clinic.util.preference.OptionalRxPreferencesConverter
 import java.net.URI
-import javax.inject.Named
+import java.time.Instant
 
-@Module(includes = [AssistedInjectModule::class])
+@Module(includes = [
+  AssistedInjectModule::class,
+  SetupActivityConfigModule::class
+])
 class SetupActivityModule {
 
   @Provides
-  @Named("fallback")
+  @TypedPreference(FallbackCountry)
   fun providesFallbackCountry(): Country {
     return Country(
         isoCountryCode = "IN",
         endpoint = URI.create(BuildConfig.FALLBACK_ENDPOINT),
         displayName = "India",
         isdCode = "91"
+    )
+  }
+
+  @Provides
+  @TypedPreference(DatabaseMaintenanceRunAt)
+  fun providesDatabaseMaintenanceRunAt(
+      rxSharedPreferences: RxSharedPreferences
+  ): Preference<Optional<Instant>> {
+    return rxSharedPreferences.getObject(
+        "database_maintenance_run_at",
+        Optional.empty(),
+        OptionalRxPreferencesConverter(InstantRxPreferencesConverter())
     )
   }
 }